Mini_project/requirements.txt
langchain
langchain-core
langchain-huggingface
langchain-community
sentence-transformers
python-dotenv
faiss-cpu
youtube-transcript-api
faster-whisper
streamlit
google-search-results
pytesseract
transformers
huggingface-hub
fastapi
uvicorn
python-multipart
reportlab